Reorganize semantic index tests into basic and patterns files
- Rename semantic_index_test.cpp to semantic_index_basic_test.cpp
- Extract complex SystemVerilog patterns to semantic_index_patterns_test.cpp
- Update BUILD.bazel with new test targets
- All 11 test cases pass across both files (9 basic + 2 pattern tests)
